fix(lambda): Fix NPE when delete operation is called on nonexistent function (#5007)

+++ clouddriver-lambda/src/main/java/com/netflix/spinnaker/clouddriver/lambda/deploy/ops/DeleteLambdaAtomicOperation.java
@@ -47,14 +47,20 @@ public class DeleteLambdaAtomicOperation
         (LambdaFunction) lambdaFunctionProvider.getFunction(account, region, functionName);
 
     AWSLambda client = getLambdaClient();
-    DeleteFunctionRequest request =
-        new DeleteFunctionRequest().withFunctionName(cache.getFunctionArn());
 
-    request.withQualifier(description.getQualifier());
+    if (cache != null && client != null) {
+      DeleteFunctionRequest request =
+          new DeleteFunctionRequest().withFunctionName(cache.getFunctionArn());
 
-    DeleteFunctionResult result = client.deleteFunction(request);
-    updateTaskStatus("Finished deletion of AWS Lambda Function  Operation...");
+      request.withQualifier(description.getQualifier());
 
-    return result;
+      DeleteFunctionResult result = client.deleteFunction(request);
+
+      updateTaskStatus("Finished deletion of AWS Lambda Function  Operation...");
+
+      return result;
+    }
+
+    return null;
   }
 }
